+++ imio/scan_helpers/tests.py
@@ -46,7 +46,7 @@ class TestUtils(unittest.TestCase):
         self.assertTrue(os.path.exists(p("test_env/download.txt")))
         with open(p("test_env/download.txt"), "r") as file:
             content = file.read()
-            self.assertIn("pyinstaller\nrequests", content)
+            self.assertIn("pip-system-certs\npyinstaller", content)
         os.remove(p("test_env/download.txt"))
 
     def test_get_dated_backup_dir(self):