+++ CHANGES.rst
@@ -1,7 +1,7 @@
 Changelog
 =========
 
-0.6.1 (unreleased)
+0.7.0 (unreleased)
 ------------------
 
 - Used `pip-system-certs` to resolve certificate problems.
@@ -10,6 +10,8 @@ Changelog
   [sgeulette]
 - Improved send_log_message to avoid timeout.
   [sgeulette]
+- Added exception handling when removing profiles directory.
+  [chris-adam]
 
 0.6.0 (2024-08-28)
 ------------------